// src/model/generated/vehicleCategoryBookingPublic.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * PASS2RENT widgets
 * For partner reservation website
 * OpenAPI spec version: v0.0.1
 */
import type { VehiclePublic } from "./vehiclePublic";
import type { BookingPlanDetail } from "./bookingPlanDetail";

export interface VehicleCategoryBookingPublic {
  uuid: string;
  name: string;
  vehicle: VehiclePublic;
  isExactVehicle?: boolean;
  bookingPlan?: BookingPlanDetail;
}